+++ R/labelled.R
@@ -13,8 +13,8 @@
 #'   of the values might be labelled.
 #' @param levels When coercing a labelled character vector to a factor, you
 #'   can choose whether to use the labels or the values as the factor levels.
-#' @param ordered If \code{TRUE} for ordinal factors, \code{FALSE} (the default)
-#'   for nominal factors.
+#' @param is_na Optionally, a logical vector describing which levels should
+#'   be translated to missing values.
 #' @param ... Ignored
 #' @export
 #' @examples
@@ -27,12 +27,20 @@
 #' as_factor(s1, labels = "values")
 #' as_factor(s2)
 #'
+#' # Other statistical software supports multiple types of missing values
+#' s3 <- labelled(c("M", "M", "F", "X", "N/A"),
+#'   c(Male = "M", Female = "F", Refused = "X", "Not applicable" = "N/A"),
+#'   c(FALSE, FALSE, TRUE, TRUE)
+#' )
+#' s3
+#' as_factor(s3)
+#'
 #' # Often when you have a partially labelled numeric vector, labelled values
 #' # are special types of missing. Use XXX to replace labels with missing
 #' # values
 #' x <- labelled(c(1, 2, 1, 2, 10, 9), c(Unknown = 9, Refused = 10))
 #' zap_labels(x)
-labelled <- function(x, labels) {
+labelled <- function(x, labels, is_na = NULL) {
   if (!is.numeric(x) && !is.character(x)) {
     stop("`x` must be either numeric or a character vector", call. = FALSE)
   }
@@ -42,9 +50,18 @@ labelled <- function(x, labels) {
   if (is.null(labels)) {
     stop("`labels` must be a named vector", call. = FALSE)
   }
+  if (is.null(is_na)) {
+    is_na <- rep(FALSE, length(labels))
+  } else {
+    if (!is.logical(is_na) || length(is_na) != length(labels)) {
+      stop("`is_na` must be a logical vector the same length as `labels`",
+        call. = FALSE)
+    }
+  }
 
   structure(x,
     labels = labels,
+    is_na = is_na,
     class = "labelled"
   )
 }
@@ -53,7 +70,7 @@ is.labelled <- function(x) inherits(x, "labelled")
 
 #' @export
 `[.labelled` <- function(x, ...) {
-  labelled(NextMethod(), attr(x, "labels"))
+  labelled(NextMethod(), attr(x, "labels"), attr(x, "is_na"))
 }
 
 #' @export
@@ -62,10 +79,13 @@ print.labelled <- function(x, ...) {
 
   xx <- unclass(x)
   attr(xx, "labels") <- NULL
-  print(xx)
+  attr(xx, "is_na") <- NULL
+  print(xx, quote = FALSE)
 
-  cat("Labels:\n")
-  print(attr(x, "labels"))
+  cat("\nLabels:\n")
+  labels <- attr(x, "labels")
+  lab_df <- data.frame(value = unname(labels), label = names(labels), is_na = attr(x, "is_na"))
+  print(lab_df, row.names = FALSE)
 
   invisible()
 }
@@ -81,22 +101,31 @@ as.data.frame.labelled <- function(x, ...) {
 }
 
 
+#' @param ordered If \code{TRUE} for ordinal factors, \code{FALSE} (the default)
+#'   for nominal factors.
+#' @param drop_na If \code{TRUE}, the default, all types are missing value are
+#'   converted into \code{NA}. If \code{FALSE}, missing values will be left as
+#'   their original codes.
 #' @rdname labelled
 #' @export
 as_factor.labelled <- function(x, levels = c("labels", "values"),
-                               ordered = FALSE, ...) {
+                               ordered = FALSE, drop_na = TRUE, ...) {
   levels <- match.arg(levels)
 
+  labels <- attr(x, "labels")
+  if (drop_na) {
+    labels <- labels[!attr(x, "is_na")]
+  }
+
   if (is.character(x)) {
-    levs <- unname(attr(x, "labels"))
+    levs <- unname(labels)
     labs <- switch(levels,
-      labels = names(attr(x, "labels")),
+      labels = names(labels),
       values = levs
     )
     factor(x, levs, labels = labs, ordered = ordered)
   } else {
-    labs <- attr(x, "labels")
-    factor(x, levels = unname(labs), labels = names(labs))
+    factor(x, levels = unname(labels), labels = names(labels))
   }
 
 }
@@ -109,6 +138,7 @@ zap_labels <- function(x) {
 
   labelled <- x %in% attr(x, "labels")
   attr(x, "labels") <- NULL
+  attr(x, "is_na") <- NULL
   class(x) <- NULL
 
   x[labelled] <- NA